+++ modules/mysql-flexible-server/locals.tf
@@ -1,2 +1,10 @@
 # locals.tf - Local values
-# (No local values defined for this module)
+
+locals {
+  default_server_configurations = {
+    "require_secure_transport" = "ON"
+    "tls_version"              = "TLSv1.2"
+  }
+
+  merged_server_configurations = merge(local.default_server_configurations, var.server_configurations)
+}