+++ modules/postgresql-flexible-server/locals.tf
@@ -1,2 +1,9 @@
 # locals.tf - Local values
-# (No local values defined for this module)
+
+locals {
+  default_server_configurations = {
+    "require_secure_transport" = "on"
+  }
+
+  merged_server_configurations = merge(local.default_server_configurations, var.server_configurations)
+}